Move nickname to a separate component
